Added test case for OrientDB. Moved collect and take extension methods from TestExtensions to SparkPimpers.
