+++ sparkplug-extras/src/main/scala/springnz/sparkplug/testkit/TestExtensions.scala
@@ -87,63 +87,4 @@ object TestExtensions {
       }
 
   }
-
-  // Converts RDDs to Types that can be verified in test assertions
-  implicit class TestResultConverter[A: ClassTag](operation: SparkOperation[RDD[A]])(implicit log: Logger) {
-    def count = {
-      operation.map { rdd ⇒
-        val n = rdd.count
-        log.debug(s"Produced an RDD of size ${n}")
-        n
-      }
-    }
-
-    def collect(): SparkOperation[Array[A]] = {
-      operation.map { rdd ⇒
-        val count = rdd.count
-        log.debug(s"Produced an RDD of size $count")
-        rdd.collect()
-      }
-    }
-
-    def take(n: Int): SparkOperation[Array[A]] = {
-      operation.map { rdd ⇒
-        val count = rdd.count
-        log.debug(s"Produced an RDD of size $count")
-        rdd.take(n)
-      }
-    }
-
-    def takeOrdered(n: Int)(implicit ord: Ordering[A]): SparkOperation[Array[A]] = {
-      operation.map { rdd ⇒
-        val count = rdd.count
-        log.debug(s"Produced an RDD of size $count")
-        rdd.takeOrdered(n)
-      }
-    }
-
-    def collectWithCount(): SparkOperation[(Array[A], Long)] = {
-      operation.map { rdd ⇒
-        val count = rdd.count
-        log.debug(s"Produced an RDD of size $count")
-        (rdd.collect, count)
-      }
-    }
-
-    def takeWithCount(n: Int): SparkOperation[(Array[A], Long)] = {
-      operation.map { rdd ⇒
-        val count = rdd.count
-        log.debug(s"Produced an RDD of size $count")
-        (rdd.take(n), count)
-      }
-    }
-
-    def takeOrderedWithCount(n: Int)(implicit ord: Ordering[A]): SparkOperation[(Array[A], Long)] = {
-      operation.map { rdd ⇒
-        val count = rdd.count
-        log.debug(s"Produced an RDD of size $count")
-        (rdd.takeOrdered(n), count)
-      }
-    }
-  }
 }
